+++ Sources/TuistDependenciesTesting/Carthage/MockCarthageCommandGenerator.swift
@@ -8,17 +8,17 @@ public final class MockCarthageCommandGenerator: CarthageCommandGenerating {
 
     var invokedCommand = false
     var invokedCommandCount = 0
-    var invokedCommandParameters: (path: AbsolutePath, platforms: Set<Platform>?)?
-    var invokedCommandParametersList = [(path: AbsolutePath, platforms: Set<Platform>?)]()
-    var commandStub: ((AbsolutePath, Set<Platform>?) -> [String])?
+    var invokedCommandParameters: (path: AbsolutePath, produceXCFrameworks: Bool, platforms: Set<Platform>?)?
+    var invokedCommandParametersList = [(path: AbsolutePath, produceXCFrameworks: Bool, platforms: Set<Platform>?)]()
+    var commandStub: ((AbsolutePath, Bool, Set<Platform>?) -> [String])?
 
-    public func command(path: AbsolutePath, platforms: Set<Platform>?) -> [String] {
+    public func command(path: AbsolutePath, produceXCFrameworks: Bool, platforms: Set<Platform>?) -> [String] {
         invokedCommand = true
         invokedCommandCount += 1
-        invokedCommandParameters = (path, platforms)
-        invokedCommandParametersList.append((path, platforms))
+        invokedCommandParameters = (path, produceXCFrameworks, platforms)
+        invokedCommandParametersList.append((path, produceXCFrameworks, platforms))
         if let stub = commandStub {
-            return stub(path, platforms)
+            return stub(path, produceXCFrameworks, platforms)
         } else {
             return []
         }