+++ Sources/TuistGenerator/GraphViz/GraphToGraphVizMapper.swift
@@ -2,61 +2,71 @@ import Foundation
 import GraphViz
 import TSCBasic
 import TuistCore
+import TuistGraph
 
 /// Interface that describes a mapper that convers a project graph into a GraphViz graph.
 protocol GraphToGraphVizMapping {
     /// Maps the project graph into a dot graph representation.
     ///
-    /// - Parameter graph: Graph to be converted into a GraphViz.Graph.
+    /// - Parameter graph: ValueGraph to be converted into a GraphViz.Graph.
     /// - Returns: The GraphViz.Graph representation.
-    func map(graph: TuistCore.Graph, skipTestTargets: Bool, skipExternalDependencies: Bool, targetsToFilter: [String]) -> GraphViz.Graph
+    func map(graph: ValueGraph, skipTestTargets: Bool, skipExternalDependencies: Bool, targetsToFilter: [String]) -> GraphViz.Graph
 }
 
 final class GraphToGraphVizMapper: GraphToGraphVizMapping {
     /// Maps the project graph into a GraphViz graph representation.
     ///
-    /// - Parameter graph: TuistCore.Graph to be converted into a GraphViz.Graph.
+    /// - Parameter graph: ValueGraph to be converted into a GraphViz.Graph.
     /// - Returns: The GraphViz.Graph representation.
-    func map(graph: TuistCore.Graph, skipTestTargets: Bool, skipExternalDependencies: Bool, targetsToFilter: [String]) -> GraphViz.Graph {
+    func map(graph: ValueGraph, skipTestTargets: Bool, skipExternalDependencies: Bool, targetsToFilter: [String]) -> GraphViz.Graph {
         var nodes: [GraphViz.Node] = []
         var dependencies: [GraphViz.Edge] = []
         var graphVizGraph = GraphViz.Graph(directed: true)
 
-        let filteredTargets = graph.targets.flatMap(\.value).filter { target in
-            if skipTestTargets, target.dependsOnXCTest {
+        let graphTraverser = ValueGraphTraverser(graph: graph)
+
+        let filteredTargets: Set<ValueGraphTarget> = graphTraverser.allTargets().filter { target in
+            if skipTestTargets, graphTraverser.dependsOnXCTest(path: target.path, name: target.target.name) {
                 return false
             }
-            if skipExternalDependencies, target.isExternal {
-                return false
-            }
-            if !targetsToFilter.isEmpty, !targetsToFilter.contains(target.name) {
+
+            if !targetsToFilter.isEmpty, !targetsToFilter.contains(target.target.name) {
                 return false
             }
 
             return true
         }
-        let filteredTargetsAndDependencies = filteredTargets + transitiveClosure(filteredTargets, successors: \.targetDependencies)
+
+        let filteredTargetsAndDependencies: Set<ValueGraphTarget> = filteredTargets.union(
+            transitiveClosure(Array(filteredTargets)) { target in
+                Array(graphTraverser.directTargetDependencies(path: target.path, name: target.target.name))
+            }
+        )
 
         filteredTargetsAndDependencies.forEach { target in
-            if skipTestTargets, target.dependsOnXCTest {
-                return
-            }
-            if skipExternalDependencies, target.isExternal {
-                return
-            }
-
             var leftNode = GraphViz.Node(target.target.name)
             leftNode.applyAttributes(attributes: target.styleAttributes)
             nodes.append(leftNode)
 
-            target.dependencies.forEach { dependency in
-                var rightNode = GraphViz.Node(dependency.name)
-                rightNode.applyAttributes(attributes: dependency.styleAttributes)
-                nodes.append(rightNode)
-                if skipExternalDependencies, dependency.isExternal { return }
-                let edge = GraphViz.Edge(from: leftNode, to: rightNode)
-                dependencies.append(edge)
-            }
+            guard
+                let targetDependencies = graphTraverser.dependencies[.target(name: target.target.name, path: target.path)]
+            else { return }
+            targetDependencies
+                .filter { dependency in
+                    if skipExternalDependencies, dependency.isExternal { return false }
+                    return true
+                }
+                .forEach { dependency in
+                    var rightNode = GraphViz.Node(dependency.name)
+                    rightNode.applyAttributes(
+                        attributes: dependency.styleAttributes(
+                            graphTraverser: graphTraverser
+                        )
+                    )
+                    nodes.append(rightNode)
+                    let edge = GraphViz.Edge(from: leftNode, to: rightNode)
+                    dependencies.append(edge)
+                }
         }
 
         let sortedNodes = Set(nodes).sorted { $0.id < $1.id }
@@ -67,27 +77,56 @@ final class GraphToGraphVizMapper: GraphToGraphVizMapping {
     }
 }
 
-private extension GraphNode {
+private extension ValueGraphDependency {
     var isExternal: Bool {
-        if self is SDKNode {
+        switch self {
+        case .target:
+            return false
+        case .framework, .xcframework, .library, .packageProduct, .sdk, .cocoapods:
             return true
         }
-        if self is CocoaPodsNode {
-            return true
-        }
-        if self is FrameworkNode {
-            return true
-        }
-        if self is LibraryNode {
-            return true
-        }
-        if self is PackageProductNode {
-            return true
-        }
-        if self is PrecompiledNode {
-            return true
-        }
+    }
 
-        return false
+    var name: String {
+        switch self {
+        case let .target(name: name, path: _):
+            return name
+        case let .framework(
+            path: path,
+            binaryPath: _,
+            dsymPath: _,
+            bcsymbolmapPaths: _,
+            linking: _,
+            architectures: _,
+            isCarthage: _
+        ):
+            return path.basenameWithoutExt
+        case let .xcframework(
+            path: path,
+            infoPlist: _,
+            primaryBinaryPath: _,
+            linking: _
+        ):
+            return path.basenameWithoutExt
+        case let .library(
+            path: path,
+            publicHeaders: _,
+            linking: _,
+            architectures: _,
+            swiftModuleMap: _
+        ):
+            return path.basenameWithoutExt
+        case let .packageProduct(path: _, product: product):
+            return product
+        case let .sdk(
+            name: name,
+            path: _,
+            status: _,
+            source: _
+        ):
+            return String(name.split(separator: ".").first ?? "")
+        case .cocoapods:
+            return "CocoaPods"
+        }
     }
 }